feat: 생성 폼 타이틀 입력 컴포넌트를 공통 컴포넌트로 생성
